refactor: children type 수정
