+++ src/components/headerContainer/index.tsx
@@ -1,6 +1,7 @@
+import { ReactNode } from "react";
 import { Container } from "./index.style";
 
-const HeaderContainer = ({ children }: { children: JSX.Element }) => {
+const HeaderContainer = ({ children }: { children: ReactNode }) => {
   return <Container>{children}</Container>;
 };
