+++ src/components/loadingContainer/index.tsx
@@ -1,6 +1,7 @@
+import { ReactNode } from "react";
 import { Container } from "./index.style";
 
-const LoadingContainer = ({ children }: { children: JSX.Element }) => {
+const LoadingContainer = ({ children }: { children: ReactNode }) => {
   return <Container>{children}</Container>;
 };
