+++ src/pages/homeLoading/index.style.tsx
